Refactor cli into UI and cmd handling script
